+++ packages/react-day-picker/src/components/HeadRow/index.ts
@@ -1,0 +1,1 @@
+export * from './HeadRow';
