+++ packages/react-day-picker/src/components/HeadRow/utils/getWeekdays.ts
